backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
requests==2.31.0
openai==0.28.0

